tweets_streaming/requirements.txt
tweepy==3.10.0
boto3==1.16.63
python-decouple==3.4